dash-app/requirements.txt
dash==2.18.2
dash_bootstrap_components==1.6.0
homeassistant_api==4.2.2.post2
plotly==5.24.1
PyYAML==6.0.2